Conflicts:
	Tendu/src/it/chalmers/tendu/Tendu.java
	Tendu/src/it/chalmers/tendu/controllers/InputController.java
	Tendu/src/it/chalmers/tendu/controllers/ShapeGameModelController.java
	Tendu/src/it/chalmers/tendu/gamemodel/GameId.java
	Tendu/src/it/chalmers/tendu/gamemodel/MiniGame.java
	Tendu/src/it/chalmers/tendu/gamemodel/numbergame/NumberGame.java
	Tendu/src/it/chalmers/tendu/gamemodel/shapesgame/ShapesGame.java
	Tendu/src/it/chalmers/tendu/screens/GameScreen.java
	Tendu/src/it/chalmers/tendu/screens/NumberGameScreen.java
	Tendu/src/it/chalmers/tendu/screens/ShapesGameScreen.java
	Tendu/src/it/chalmers/tendu/tbd/C.java

+++ Tendu/src/it/chalmers/tendu/gamemodel/MiniGameFactory.java
@@ -21,7 +21,7 @@ public class MiniGameFactory {
 		return game;
 	}
 
-	public static MiniGame createMiniGame(int bonusTime, GameId gameId,
+	public static MiniGame createMiniGame(long bonusTime, GameId gameId,
 			Difficulty difficulty, Map<String, Integer> players) {
 
 		MiniGame miniGame = null;
